jmap_ical: reintroduce X-JMAP-USEDEFAULTALERTS property
In an earlier implementation of JSCalendar we deprecated the
X-JMAP-USEDEFAULTALERTS property in favor of the custom Apple
property X-APPLE-DEFAULT-ALARM.

This was in line with the JSCalendar definition, where
useDefaultAlerts defaults to "false". Since Apple's X-property
only allows the TRUE value, we accordingly set this property
only if a client explicitly set useDefaultAlerts to "true".
The "false" value was mapped by the absence of the X-property.

Now, the JMAP Calendars spec requires the useDefaultAlerts
value default to "true" for sharees of a calendar in team-mode.
This requires to reintroduce X-JMAP-USEDEFAULTALERTS, as our
JMAP X-property allows to set value FALSE. This is required
so that sharees can explicitly opt out of default alerts on
shared.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>